Refactored old code (different files)
